--- a/node.go
+++ b/node.go
@@ -214,7 +214,7 @@ func (n *skippedNodes) pop() skippedNode {
 }
 
 type skippedNode struct {
-	node      *node
+	parent    *node
 	pathIndex int
 	paramCnt  uint32
 	seen      bool