--- a/options.go
+++ b/options.go
@@ -113,10 +113,22 @@ func WithAutoOptions(enable bool) Option {
 // WithRedirectTrailingSlash enable automatic redirection fallback when the current request does not match but
 // another handler is found with/without an additional trailing slash. E.g. /foo/bar/ request does not match
 // but /foo/bar would match. The client is redirected with a http status code 301 for GET requests and 308 for
-// all other methods.
+// all other methods. Note that this option is mutually exclusive with WithIgnoreTrailingSlash, and if both are
+// enabled, WithRedirectTrailingSlash takes precedence.
 func WithRedirectTrailingSlash(enable bool) Option {
 	return optionFunc(func(r *Router) {
 		r.redirectTrailingSlash = enable
+	})
+}
+
+// WithIgnoreTrailingSlash allows the router to match routes regardless of whether a trailing slash is present or not.
+// E.g. /foo/bar/ and /foo/bar would both match the same handler. This option prevents the router from issuing
+// a redirect and instead matches the request directly. Note that this option is mutually exclusive with
+// WithRedirectTrailingSlash, and if both are enabled, WithRedirectTrailingSlash takes precedence.
+// This api is EXPERIMENTAL and is likely to change in future release.
+func WithIgnoreTrailingSlash(enable bool) Option {
+	return optionFunc(func(r *Router) {
+		r.ignoreTrailingSlash = enable
 	})
 }