… README.md
… fox_test.go
… go.sum